+++ front-page-april-start.php
@@ -119,11 +119,11 @@
                         <div class="block-container">
                             <a href="<?php echo site_url('/product-category/brands/quoizel-lighting'); ?>">
                                 <div class="block-picture">
-                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/06/quoizel_bath_sale.jpg">', wp_get_attachment_metadata(172776), 172776 ); ?>
+                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="https://lightyourhome.com/wp-content/uploads/2020/06/quoizel_bath_sale.jpg">', wp_get_attachment_metadata(172776), 172776 ); ?>
                                 </div>
                                 <div class="block-caption-wrapper">
                                     <div class="block-caption-container">
-                                        <h4 class="block-heading">Quoizel Lighting</h4>
+                                        <4 class="block-heading">Quoizel Lighting</h4>
                                         <h3 class="block-heading">Save 10% on Quoizel</h3>
                                         <h4 class="block-heading">Use code "QUOIZEL10" at checkout!</h4>
                                         <a style="text-align: center;" href="<?php echo site_url('/product-category/brands/quoizel-lighting'); ?>" class="block-shop-link">Shop Now</a>
@@ -135,19 +135,19 @@
 
                      <div class="col-md-6">
                         <div class="block-container">
-                            <a href="<?php echo site_url('/brands/kichler-lighting'); ?>">
+                            <a href="<?php echo site_url('/brands/elk-lighting'); ?>">
                                 <div class="block-picture">
-                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/06/kichler_summer_sale_2020.jpg">', wp_get_attachment_metadata(172667), 172667 ); ?>
+                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="https://lightyourhome.com/wp-content/uploads/2020/05/89047_8.jpg">', wp_get_attachment_metadata(166726), 166726 ); ?>
                                 </div>
                                 <div class="block-caption-wrapper">
                                     <div class="block-caption-container">
-                                        <h4 class="block-heading">Kichler Lighting</h4>
-                                        <h3 class="block-heading">Save 20% off all Kichler Lighting</h3>
-                                        <h4 class="block-heading">Prices already discounted!</h4>
-                                        <a style="text-align: center;" href="<?php echo site_url('/brands/kichler-lighting'); ?>" class="block-shop-link">Shop Now</a>
+                                        <h4 class="block-heading">Elk Lighting</h4>
+                                        <h3 class="block-heading">Save 15% off all Elk Lighting</h3>
+                                        <h4 class="block-heading">Use code "ELKLIGHTING15" at checkout!</h4>
+                                        <a style="text-align: center;" href="<?php echo site_url('/brands/elk-lighting'); ?>" class="block-shop-link">Shop Now</a>
                                     </div>
                                 </div>
-                            </a>
+                             </a>
                         </div>
                      </div>