+++ sale-page-et2-maxim-update.php
@@ -88,42 +88,6 @@
                          </div>
                      </div>
 
-                     <div class="col-md-4">
-                        <div class="block-container">
-                            <a href="<?php echo site_url('/brands/kichler-lighting'); ?>">
-                                <div class="block-picture">
-                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/07/43036DAG-10.jpg">', wp_get_attachment_metadata(29834), 29834 ); ?>
-                                </div>
-                                <div class="block-caption-wrapper">
-                                    <div class="block-caption-container">
-                                        <h4 class="block-heading">Kichler Lighting</h4>
-                                        <h3 class="block-heading">Save 20% off all Kichler Lighting</h3>
-                                        <h4 class="block-heading">Prices already discounted!</h4>
-                                        <a style="text-align: center;" href="<?php echo site_url('/brands/kichler-lighting'); ?>" class="block-shop-link">Shop Now</a>
-                                    </div>
-                                </div>
-                             </a>
-                        </div>
-                     </div>
-                     
-                     <div class="col-md-4">
-                        <div class="block-container">
-                            <a href="<?php echo site_url('/brands/elan-lighting'); ?>">
-                                <div class="block-picture">
-                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/01/84195.jpg">', wp_get_attachment_metadata(106870), 106870 ); ?>
-                                </div>
-                                <div class="block-caption-wrapper">
-                                    <div class="block-caption-container">
-                                        <h4 class="block-heading">Élan Lighting</h4>
-                                        <h3 class="block-heading">Save 20% off all Élan Lighting</h3>
-                                        <h4 class="block-heading">Prices already discounted!</h4>
-                                        <a style="text-align: center;" href="<?php echo site_url('/brands/elan-lighting'); ?>" class="block-shop-link">Shop Now</a>
-                                    </div>
-                                </div>
-                             </a>
-                        </div>
-                     </div>
-
                      <!-- <div class="col-md-4">
                         <div class="block-container">
                             <a href="<?php //echo site_url('/brands/innovations-lighting'); ?>">